refactor: pox lock handling into separate workspace member
