+++ pox-locking/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "pox-locking"
+version = "2.4.0"
+authors = [ "Jude Nelson <jude@stacks.org>",
+            "Aaron Blankstein <aaron@blockstack.com>",
+            "Ludo Galabru <ludovic@blockstack.com>" ]
+license = "GPLv3"
+homepage = "https://github.com/blockstack/stacks-blockchain"
+repository = "https://github.com/blockstack/stacks-blockchain"
+description = "Contract call result handler for applying PoX lock operations"
+keywords = [ "stacks", "stx", "bitcoin", "crypto", "blockstack", "decentralized", "dapps", "blockchain" ]
+readme = "README.md"
+resolver = "2"
+edition = "2021"
+rust-version = "1.61"
+
+[lib]
+name = "pox_locking"
+path = "src/lib.rs"
+
+[dependencies]
+clarity = { package = "clarity", path = "../clarity/." }
+stacks_common = { package = "stacks-common", path = "../stacks-common/." }
+slog = { version = "2.5.2", features = [ "max_level_trace" ] }
+
+[features]
+slog_json = ["stacks_common/slog_json", "clarity/slog_json"]
